TimeSeries
The graph shows the Proportion of seats held by women in national
parliaments (%). The purpose of doing this is women in general are not
actively involved in politics and people talk of equality of status
everywhere. The graphs shows the conditions. link to dataset
http://data.worldbank.org/indicator/SG.GEN.PARL.ZS I have not used the
exact dataset but picked few countries which I feel are relevant and few
years of data and changed the format to work with dimple.